+++ module/base_decorator.rb
@@ -1,0 +1,14 @@
+require '../school_library'
+require './module/nameable'
+class BaseDecorator < Nameable
+  attr_accessor :nameable
+
+  def initialize(nameable)
+    super()
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable
+  end
+end
