+++ module/capitalize_decorator.rb
@@ -1,0 +1,6 @@
+require './base_decorator'
+class CapitalizeDecorator < BaseDecorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
