+++ school_library.rb
@@ -1,5 +1,7 @@
 require './module/nameable'
 class Person < Nameable
+  attr_accessor :name, :age
+
   def initialize(age, name = 'Unknown', parent_permission: true)
     super()
     @id = Random.rand(1..100)
@@ -7,11 +9,6 @@ class Person < Nameable
     @age = age
     @parent_permission = parent_permission
   end
-
-  # Getters
-  attr_accessor :name
-
-  attr_accessor :age
 
   def can_use_services
     return unless @age >= 18 || parent_permission
@@ -27,30 +24,5 @@ class Person < Nameable
 
   def of_age
     @age >= 18
-  end
-end
-
-class BaseDecorator < Nameable
-  attr_accessor :nameable
-
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable
-  end
-end
-
-class CapitalizeDecorator < BaseDecorator
-  def correct_name
-    @nameable.upcase!
-  end
-end
-
-class TrimmerDecorator < BaseDecorator
-  def correct_name
-    @nameable.splice(0, 9)
   end
 end
